+++ js/utils.js
@@ -1,0 +1,3 @@
+export function checkString(str, maxLength) {
+    return str.length < maxLength;
+}
