+++ package.json
@@ -2,6 +2,7 @@
   "name": "kekstagram",
   "version": "26.0.0",
   "private": true,
+  "type": "module",
   "description": "Личный проект «Кекстаграм» от HTML Academy",
   "repository": {
     "type": "git",
@@ -25,3 +26,5 @@
     "npm": "8.x"
   }
 }
+
+
