+++ ts/main.ts
@@ -1,21 +1,48 @@
+import { commentTexts, names } from "./data/names.js";
+import { IPhotoDescription, IComment } from "./interfaces.js";
 import { getPhotoDescriptions } from "./photo.js";
 import { getRandomNumber, getUniqueRandomNumber } from "./random.js";
 import { checkString } from "./utils.js";
+
 console.log(1);
-const randomCount = 25;
-const maxLenght = 10;
-const str = 'В этой строке точно больше 10 символов';
-const min = 12;
-const max = 20;
+
+const randomCount: number = 25;
+const maxLenght: number = 10;
+const str: string = 'В этой строке точно больше 10 символов';
+const min: number = 12;
+const max: number = 20;
+
 const photoDescriptions = getPhotoDescriptions(randomCount);
 console.log(photoDescriptions);
-const isCorrectly = checkString(str, maxLenght);
+
+const isCorrectly: boolean = checkString(str, maxLenght);
 console.log(isCorrectly);
-const randomNumber = getRandomNumber(min, max);
+
+const randomNumber: number = getRandomNumber(min, max);
 console.log('Случайное число от ' + min + ' до ' + max + ' = ' + randomNumber);
-const uniqueRandomNumbers = [];
+
+const uniqueRandomNumbers: number[] = [];
 for (let i = 0; i < 7; i++) {
-    const uniqueRandomNumber = getUniqueRandomNumber(min, max);
+    const uniqueRandomNumber: number = getUniqueRandomNumber(min, max);
     uniqueRandomNumbers.push(uniqueRandomNumber);
 }
 console.log('Случайные и уникальные числа от ' + min + ' до ' + max + ' = ' + uniqueRandomNumbers);
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
