+++ ts/utils.ts
@@ -1,0 +1,4 @@
+
+export function checkString(str: string, maxLength: number): boolean {
+    return str.length < maxLength;
+}
